pyproject.toml
[project]
name = "tau-anomaly-detection"
version = "0.1.0"
description = "Anomaly detection in ATLAS tau data using autoencoders and variational autoencoders"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "atlas-mpl-style>=0.25.0",
    "awkward>=2.9.0",
    "fastapi>=0.115.0",
    "hydra-core>=1.3.2",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "lightning>=2.5.0",
    "matplotlib>=3.10.8",
    "mlflow>=3.9.0",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "pandera>=0.23.0",
    "pyrootutils>=1.0.4",
    "ray[tune]>=2.40.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "torcheval>=0.0.7",
    "torchmetrics>=1.7.0",
    "tqdm>=4.67.3",
    "uproot>=5.7.1",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.15.2",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["src/__init__.py", "src/*/__init__.py"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    "if __name__ == .__main__.",
    "pragma: no cover",
]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
warn_unused_ignores = true
no_implicit_optional = true

# Strict checking for serving layer
[[tool.mypy.overrides]]
module = ["src.serving.*"]
disallow_untyped_defs = true
warn_return_any = true

# Hydra OmegaConf config access patterns
[[tool.mypy.overrides]]
module = [
    "src.processing.merger",
    "src.processing.features",
]
disable_error_code = ["union-attr", "arg-type", "index", "call-overload", "assignment"]

# Ray Tune + Lightning interactions
[[tool.mypy.overrides]]
module = ["src.models.tuning"]
disable_error_code = ["union-attr", "arg-type", "index", "call-overload", "assignment", "call-arg"]

# Matplotlib type stubs stricter than runtime API
[[tool.mypy.overrides]]
module = [
    "src.models.evaluation",
    "src.models.plots",
    "src.eda.plots",
    "src.visualization.*",
]
disable_error_code = ["arg-type", "attr-defined", "index", "union-attr"]

[tool.ruff]
extend-exclude = ["notebooks/*.ipynb"]
